Reorganize nullability tests in :build-logic:binary-compatibility and add failing JSpecify coverage
