+++ build-logic/binary-compatibility/src/test/kotlin/gradlebuild/binarycompatibility/AbstractBinaryCompatibilityTest.kt
@@ -113,7 +113,6 @@ abstract class AbstractBinaryCompatibilityTest {
                     package com.example
 
                     import org.gradle.api.Incubating
-                    import javax.annotation.Nullable
 
                     $v1
                     """
@@ -126,7 +125,6 @@ abstract class AbstractBinaryCompatibilityTest {
                     package com.example
 
                     import org.gradle.api.Incubating
-                    import javax.annotation.Nullable
 
                     $v2
                     """
@@ -145,7 +143,6 @@ abstract class AbstractBinaryCompatibilityTest {
                     package com.example;
 
                     import org.gradle.api.Incubating;
-                    import javax.annotation.Nullable;
 
                     $v1
                     """
@@ -158,7 +155,6 @@ abstract class AbstractBinaryCompatibilityTest {
                     package com.example;
 
                     import org.gradle.api.Incubating;
-                    import javax.annotation.Nullable;
 
                     $v2
                     """
@@ -271,6 +267,8 @@ abstract class AbstractBinaryCompatibilityTest {
                         dependencies {
                             "implementation"(gradleApi())
                             "implementation"(kotlin("stdlib"))
+                            // TODO remove once JSpecify is part of the Gradle API
+                            "implementation"("org.jspecify:jspecify:1.0.0")
                         }
                     }
                     project(":v1") {
